Add switch to response/code editor
