+++ Composer/packages/lib/code-editor/src/lg/LgResponseEditor.tsx
@@ -1,0 +1,13 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+import * as React from 'react';
+
+type Props = {};
+
+// This part is being implemented and tracked by task item:
+// https://github.com/microsoft/BotFramework-Composer/issues/5571
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+export const LgResponseEditor = (_: Props) => {
+  return <div>Under construction</div>;
+};
